+++ crates/zng-view/src/image_cache/encode.rs
@@ -333,9 +333,9 @@ impl Image {
                 } else {
                     image::ColorType::Rgba8.into()
                 };
-                let img = image::codecs::tga::TgaEncoder::new(buffer);                
+                let img = image::codecs::tga::TgaEncoder::new(buffer);
                 img.encode(&buf, width, height, ct)?;
-            },
+            }
             #[cfg(feature = "image_bmp")]
             image::ImageFormat::Bmp => {
                 let ct = if is_mask {
@@ -352,7 +352,7 @@ impl Image {
                     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
                         self.0.write(buf)
                     }
-                
+
                     fn flush(&mut self) -> std::io::Result<()> {
                         self.0.flush()
                     }
@@ -360,40 +360,42 @@ impl Image {
                 let mut buffer = SizedProxy(buffer);
                 let mut img = image::codecs::bmp::BmpEncoder::new(&mut buffer);
                 img.encode(&buf, width, height, ct)?;
-            },
+            }
             #[cfg(feature = "image_ico")]
             image::ImageFormat::Ico => {
                 Self::encode_ico(*size, is_mask, pixels, is_opaque, entries, buffer).map_err(|e| {
                     image::ImageError::Encoding(image::error::EncodingError::new(image::error::ImageFormatHint::Exact(format), e))
                 })?;
             }
-             #[cfg(feature = "image_hdr")]
+            #[cfg(feature = "image_hdr")]
             image::ImageFormat::Hdr => {
+                const F: f32 = 1.0 / 255.0;
                 let mut rgb = IpcBytesMutCast::<[f32; 3]>::new_blocking(width as usize * height as usize)?;
                 if is_mask {
                     for (c, a) in rgb.iter_mut().zip(buf.iter()) {
-                        let a = *a as f32 / 255.0;
+                        let a = *a as f32 * F;
                         c[0] = a;
                         c[1] = a;
                         c[2] = a;
                     }
                 } else {
                     for (c32, c8) in rgb.iter_mut().zip(buf.chunks_exact(4)) {
-                        for (c, a) in c32.iter_mut().zip(c8.iter()) {
-                            *c = *a as f32 / 255.0;
+                        for (c32, c8) in c32.iter_mut().zip(c8.iter()) {
+                            *c32 = *c8 as f32 * F;
                         }
                     }
                 }
                 let img = image::codecs::hdr::HdrEncoder::new(buffer);
                 img.write_image(rgb.as_bytes(), width, height, image::ColorType::Rgb32F.into())?;
-            },
-             #[cfg(feature = "image_exr")]
+            }
+            #[cfg(feature = "image_exr")]
             image::ImageFormat::OpenExr => {
+                const F: f32 = 1.0 / 255.0;
                 let img = image::codecs::openexr::OpenExrEncoder::new(buffer);
                 let ct = if is_mask {
                     let mut rgb = IpcBytesMutCast::<[f32; 3]>::new_blocking(width as usize * height as usize)?;
                     for (c, a) in rgb.iter_mut().zip(buf.iter()) {
-                        let a = *a as f32 / 255.0;
+                        let a = *a as f32 * F;
                         c[0] = a;
                         c[1] = a;
                         c[2] = a;
@@ -404,7 +406,7 @@ impl Image {
                     let mut rgb = IpcBytesMutCast::<[f32; 3]>::new_blocking(width as usize * height as usize)?;
                     for (c32, c8) in rgb.iter_mut().zip(buf.chunks_exact(4)) {
                         for (c, a) in c32.iter_mut().zip(c8.iter()) {
-                            *c = *a as f32 / 255.0;
+                            *c = *a as f32 * F;
                         }
                     }
                     buf = rgb.into_inner();
@@ -412,18 +414,62 @@ impl Image {
                 } else {
                     let mut rgba = IpcBytesMutCast::<[f32; 4]>::new_blocking(width as usize * height as usize)?;
                     for (c32, c8) in rgba.iter_mut().zip(buf.chunks_exact(4)) {
-                        for (c, a) in c32.iter_mut().zip(c8.iter()) {
-                            *c = *a as f32 / 255.0;
+                        for (c32, c8) in c32.iter_mut().zip(c8.iter()) {
+                            *c32 = *c8 as f32 * F;
                         }
                     }
                     buf = rgba.into_inner();
                     image::ColorType::Rgba32F.into()
                 };
                 img.write_image(&buf, width, height, ct)?;
-            },
-            image::ImageFormat::Farbfeld => todo!(),
-            image::ImageFormat::Avif => todo!(),
-            image::ImageFormat::Qoi => todo!(),
+            }
+            #[cfg(feature = "image_ff")]
+            image::ImageFormat::Farbfeld => {
+                const F: u16 = 257;
+                let mut rgba = IpcBytesMutCast::<[u16; 4]>::new_blocking(width as usize * height as usize)?;
+                if is_mask {
+                    for (c, a) in rgba.iter_mut().zip(buf.iter()) {
+                        let a = *a as u16 * 257;
+                        c[0] = a;
+                        c[1] = a;
+                        c[2] = a;
+                        c[3] = u16::MAX;
+                    }
+                } else {
+                    for (c16, c8) in rgba.iter_mut().zip(buf.chunks_exact(4)) {
+                        for (c16, c8) in c16.iter_mut().zip(c8.iter()) {
+                            *c16 = *c8 as u16 * F;
+                        }
+                    }
+                }
+                let img = image::codecs::farbfeld::FarbfeldEncoder::new(buffer);
+                img.encode(rgba.as_bytes(), width, height)?;
+            }
+            #[cfg(feature = "image_qoi")]
+            image::ImageFormat::Qoi => {
+                if is_mask {
+                    // encoder only supports RGB
+                    let mut expanded = IpcBytesMut::new_blocking(buf.len() * 3)?;
+                    for (p, &a) in expanded.chunks_exact_mut(3).zip(&buf[..]) {
+                        p[0] = a;
+                        p[1] = a;
+                        p[2] = a;
+                    }
+                } else if is_opaque {
+                    buf.reduce_in_place(|[r, g, b, _]| [r, g, b]);
+                };
+                let img = image::codecs::qoi::QoiEncoder::new(buffer);
+                img.write_image(
+                    &buf,
+                    width,
+                    height,
+                    if is_opaque || is_mask {
+                        image::ColorType::Rgb8.into()
+                    } else {
+                        image::ColorType::Rgba8.into()
+                    },
+                )?;
+            }
             f => {
                 return Err(image::ImageError::Encoding(image::error::EncodingError::new(
                     image::error::ImageFormatHint::Exact(f),